2d attempt
