+++ frontend/src/components/cube.css
@@ -1,43 +1,66 @@
-.perspective {
-    perspective: 3000px;
-  }
-  
-  .cube {
-    width: 12rem;
-    height: 12rem;
-    position: relative;
-    transform-style: preserve-3d;
-    transform: rotateX(-30deg) rotateY(-30deg);
-  }
-  
-  .face {
-    position: absolute;
-    width: 12rem;
-    height: 12rem;
-    border: 1px solid black;
-  }
-  
-  .front {
-    transform: rotateY(0deg) translateZ(6rem);
-  }
-  
-  .back {
-    transform: rotateY(180deg) translateZ(6rem);
-  }
-  
-  .left {
-    transform: rotateY(-90deg) translateZ(6rem);
-  }
-  
-  .right {
-    transform: rotateY(90deg) translateZ(6rem);
-  }
-  
-  .top {
-    transform: rotateX(90deg) translateZ(6rem);
-  }
-  
-  .bottom {
-    transform: rotateX(-90deg) translateZ(6rem);
-  }
-  
+.cube {
+  position: relative;
+  transform-style: preserve-3d;
+  transform: rotateX(-30deg) rotateY(-30deg);
+}
+
+.face {
+  position: absolute;
+  width: 12rem;
+  height: 12rem;
+  border: 1px solid black;
+}
+
+.front {
+  transform: rotateY(0deg) translateZ(6rem);
+}
+
+.back {
+  transform: rotateY(180deg) translateZ(6rem);
+}
+
+.left {
+  transform: rotateY(-90deg) translateZ(6rem);
+}
+
+.right {
+  transform: rotateY(90deg) translateZ(6rem);
+}
+
+.top {
+  transform: rotateX(90deg) translateZ(6rem);
+}
+
+.bottom {
+  transform: rotateX(-90deg) translateZ(6rem);
+}
+
+.cube2d {
+  gap: 0;
+}
+
+.face2d {
+  width: 6rem;
+  height: 6rem;
+}
+
+.front2d {
+  translate:0 6rem;
+}
+
+.back2d {
+  translate: 12rem;
+}
+
+.left2d {
+translate:-6rem -6rem}
+
+.right2d {
+translate: 6rem -12rem;}
+
+.top2d {
+translate:0 -24rem
+}
+
+.bottom2d {
+translate:0 -18rem}
